Adds Order class
The Order class represents any combination
or combinations of menu items whose sum equals
the target price.

All combinations, including duplicates, of menu items
are found with Array#repeated_combination. If the sum of a combination
is equal to the target price then that combination is added to the order.

The maximum qty of a combination of items is determined by dividing
the target price by the minimum menu item price as any qty of items
in a combination greater than that will be greater than the target price.